Add routing tests
